Quebrando arquivos math.hpp e vec.h
