+++ include/math/axis_rot.hpp
@@ -1,0 +1,10 @@
+#ifndef S3D_MATH_AXIS_ROT_H
+#define S3D_MATH_AXIS_ROT_H
+
+#include "cppfn.h"
+#include "matrix/type.h"
+#include "vector/type.h"
+
+cppfn void s3d_axis_rot(s3d_fmat4 out, float angle, const s3d_fvec3 axis);
+
+#endif
