+++ include/math/euler_rot.h
@@ -1,0 +1,9 @@
+#ifndef S3D_MATH_EULER_ROT_H
+#define S3D_MATH_EULER_ROT_H
+
+#include "matrix/type.h"
+#include "vector/type.h"
+
+void s3d_euler_rot(s3d_fmat4 mat, s3d_fvec3 rot);
+
+#endif
