+++ include/math/scale.hpp
@@ -1,0 +1,10 @@
+#ifndef S3D_MATH_SCALE_H
+#define S3D_MATH_SCALE_H
+
+#include "cppfn.h"
+#include "matrix/type.h"
+#include "vector/type.h"
+
+cppfn void s3d_scale(s3d_fmat4 mat, const s3d_fvec3 val);
+
+#endif
