+++ include/math/translate.hpp
@@ -1,0 +1,10 @@
+#ifndef S3D_MATH_TRANSLATE_HPP
+#define S3D_MATH_TRANSLATE_HPP
+
+#include "cppfn.h"
+#include "matrix/type.h"
+#include "vector/type.h"
+
+cppfn void s3d_translate(s3d_fmat4 mat, const s3d_fvec3 val);
+
+#endif
