+++ include/matrix/multiply.hpp
@@ -1,0 +1,9 @@
+#ifndef S3D_MATRIX_MULTIPLY_H
+#define S3D_MATRIX_MULTIPLY_H
+ 
+#include "cppfn.h"
+#include "matrix/type.h"
+
+cppfn void s3d_mat_mul(s3d_fmat4 a, const s3d_fmat4 b);
+
+#endif
