+++ include/vector/axis.h
@@ -1,0 +1,12 @@
+#ifndef S3D_VECTOR_AXIS_H
+#define S3D_VECTOR_AXIS_H
+
+#include "vector/type.h"
+
+#define s3d_axis_x (s3d_axis[0])
+#define s3d_axis_y (s3d_axis[1])
+#define s3d_axis_z (s3d_axis[2])
+
+extern s3d_fvec3 s3d_axis[3];
+
+#endif
