+++ include/vector/type.h
@@ -1,0 +1,12 @@
+#ifndef S3D_VECTOR_TYPE_H
+#define S3D_VECTOR_TYPE_H
+
+typedef int s3d_ivec2[2];
+typedef int s3d_ivec3[3];
+typedef int s3d_ivec4[4];
+
+typedef float s3d_fvec2[2];
+typedef float s3d_fvec3[3];
+typedef float s3d_fvec4[4];
+
+#endif
