+++ src/vector/axis.c
@@ -1,4 +1,6 @@
-#include "vec.h"
+#include "vector/axis.h"
+
+#include "vector/type.h"
 
 s3d_fvec3 s3d_axis[] = {
     1, 0, 0,
